requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
bcrypt


certifi
charset-normalizer
idna
python-dotenv
requests
urllib3

